Fix structure URL handling
Allow ':page:3' to work using Structure in certain situations. Previously,
the regex generated for something like

    :page:3/:any/:pagination

would look like

    #^(news/)/([^/]+)((?:/P\d+)?)$#

and would never match because there are two slashes in a row. For whatever
reason, $this->pageUris has trailing slashes on all of the items, so those
trailing slashes need to be removed when creating the rule.